refactor: improve code health
