+++ Git2SemVer.MSBuild/Versioning/VersionGeneratorFactory.cs
@@ -1,4 +1,5 @@
-﻿using NoeticTools.Common.Logging;
+﻿using Microsoft.Build.Utilities;
+using NoeticTools.Common.Logging;
 using NoeticTools.Common.Tools.Git;
 using NoeticTools.Git2SemVer.MSBuild.Framework.Config;
 using NoeticTools.Git2SemVer.MSBuild.Tools.CI;
@@ -8,24 +9,29 @@ using NoeticTools.Git2SemVer.MSBuild.Versioning.Generation.Builders.Scripting;
 using NoeticTools.Git2SemVer.MSBuild.Versioning.Persistence;
 
 
-namespace NoeticTools.Git2SemVer.MSBuild.Tasks;
+namespace NoeticTools.Git2SemVer.MSBuild.Versioning;
 
-public sealed class GenerateVersionTask
+internal sealed class VersionGeneratorFactory
 {
     private readonly ILogger _logger;
 
-    public GenerateVersionTask(ILogger logger)
+    public VersionGeneratorFactory(ILogger logger)
     {
         _logger = logger;
     }
 
-    public IVersionOutputs GenerateVersions(IVersionGeneratorInputs inputs)
+    public VersionGenerator Create(IVersionGeneratorInputs inputs)
     {
+        if (inputs == null)
+        {
+            throw new ArgumentNullException(nameof(inputs), "Inputs is required.");
+        }
+
         var config = Git2SemVerConfiguration.Load();
         var host = new BuildHostFactory(config, _logger).Create(inputs.HostType,
-                                                               inputs.BuildNumber,
-                                                               inputs.BuildContext,
-                                                               inputs.BuildIdFormat);
+                                                                inputs.BuildNumber,
+                                                                inputs.BuildContext,
+                                                                inputs.BuildIdFormat);
         var gitTool = new GitTool(_logger)
         {
             WorkingDirectory = inputs.WorkingDirectory
@@ -40,6 +46,6 @@ public sealed class GenerateVersionTask
                                                     new GeneratedVersionsPropsFile(),
                                                     gitTool, gitPathsFinder, defaultBuilderFactory,
                                                     scriptBuilder, _logger);
-        return versionGenerator.Run();
+        return versionGenerator;
     }
 }
